+++ packages/ui/src/utils/index.ts
@@ -1,3 +1,5 @@
+export { convertRoundnessMatrix } from './roundness';
+export type { PinProps } from './roundness';
 export { detectDevice, deviceScales, isSberBox, isSberPortal } from './deviceDetection';
 export type { DeviceKind } from './deviceDetection';
 export { mediaQuery, breakpoints } from './mediaQuery';
